requirements.txt
pymongo>=4.6.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
streamlit>=1.28.0
voyageai>=0.2.0
